完成图片检索VisualEntityExtract
