+++ examples/visual_entity_extract_learn.py
@@ -1,4 +1,3 @@
-# from openks.loaders import loader_config, SourceType, FileType, Loader
 from openks.models import OpenKSModel
 
 # 列出已加载模型
@@ -6,12 +5,11 @@ OpenKSModel.list_modules()
 
 # 算法模型选择配置
 args = {
-        "MODEL.DEVICE": 'cpu',
-        "mode": "entity",
+        "MODEL.DEVICE": 'cpu'
     }
 platform = 'PyTorch'
-executor = 'VisualRelation'
-model = 'pytorch-VisualEntityExtractor'
+executor = 'VisualEntityExtract'
+model = 'pytorch-VisualEntityExtractExtractor'
 print("根据配置，使用 {} 框架，{} 执行器训练 {} 模型。".format(platform, executor, model))
 print("-----------------------------------------------")
 # 模型训练
